requirements/web.txt
alembic
email-validator
fastapi
git+https://github.com/mfreeborn/fastapi-sqlalchemy@master
hiredis
psycopg2-binary
pwdlib[bcrypt]
pydantic-settings
redis
sqlalchemy
uvicorn[standard]
langchain
langchain-community
langchain-chroma
langchain-openai
langchain-ollama
duckduckgo-search
wikipedia
beautifulsoup4
lxml[html_clean]
fake_useragent
python-ulid
